Feat : 브래킬 임시 추가 & 방황 상태 및 기능 템플릿화
