+++ Assets/Scripts/Enemy/Brakill/BrakillData.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+
+namespace HaewolWorkshop
+{
+    [CreateAssetMenu(fileName = "Brakill Data", menuName = "Data/Enemy Data/Brakill Data", order = 0)]
+    public class BrakillData : EnemyData
+    {
+        
+    }
+}
