+++ Assets/Scripts/Enemy/Brakill/BrakillData.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 7a863c1e927f4abeb70dcee00a3b94e7
+timeCreated: 1678539061
