+++ Assets/Scripts/Enemy/Brakill/BrakillStates/BrakillIdleState.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+
+namespace HaewolWorkshop
+{
+    [FSMState((int)Brakill.BrakillStates.Idle)]
+    public class BrakillIdleState : MonoBehaviour
+    {
+        
+    }
+}
