+++ Assets/Scripts/Enemy/Brakill/BrakillStates/BrakillIdleState.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a25ff22a99164c72a9fb71c42a33ab6d
+timeCreated: 1678539231
