+++ Assets/Scripts/Enemy/Brakill/BrakillStates/BrakillWanderState.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace HaewolWorkshop
+{
+    [FSMState((int)Brakill.BrakillStates.Wander)]
+    public class BrakillWanderState : EnemyWanderStateTemplate<Brakill, BrakillData>
+    {
+        public BrakillWanderState(IFSMEntity entity) : base(entity)
+        {
+        }
+    }
+}
