+++ Assets/Scripts/Enemy/Brakill/BrakillStates/BrakillWanderState.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 65b06854a42b44cba1646f4b37078c7a
+timeCreated: 1678539260
