+++ Assets/Scripts/Enemy/EnemyWanderStateTemplate.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: c917d80c9fbc4f8ca26f5c5c5a128063
+timeCreated: 1678539297
