+++ Assets/Scripts/Enemy/IWanderable.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a08c1cae6b4b4fad938a6f1614106c48
+timeCreated: 1678539491
